added more functionallity

+++ app/src/main/java/com/covid/covidapps/datasource/Patient.kt
@@ -2,7 +2,7 @@ package com.covid.covidapps.datasource
 
 data class Patient(
     val FSR: Long,
-    val heartRate: Long,
+    val heartRate: Double,
     val spO2: Long,
     val temperature: Double
 )
